Cleaned Up
I cleaned things up and put in lots of comments and explanations for each function.  It still needs to be fixed so that I get the correct series of numbers, but it still works reasonably well.  Right now, with a list of integers from 1 - 100, I get a series of 68 numbers, starting with 18, but it should be 77 long and starts with 53 or something.